# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: ChaloTogether API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rides/create:
    post:
      operationId: createRide
      tags: [rides]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RideInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideDetail"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rides/search:
    get:
      operationId: searchRides
      tags: [rides]
      parameters:
        - name: origin
          in: query
          schema:
            type: string
        - name: destination
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
        - name: time
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Matching rides with scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RideSearchResult"
...
  /rides/{id}:
    get:
      operationId: getRide
      tags: [rides]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ride detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rides/{id}/book:
    post:
      operationId: bookRide
      tags: [rides]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rides/{id}/bookings:
    get:
      operationId: getRideBookings
      tags: [rides]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking requests for driver
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingWithPassenger"
  /rides/{id}/cancel:
    post:
      operationId: cancelRide
      tags: [rides]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /rides/{id}/location:
    post:
      operationId: updateRideLocation
      tags: [rides]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Bookings ────────────────────────────────────────────────────────────────
...
  /trusted-contacts:
    get:
      operationId: listTrustedContacts
      tags: [trusted-contacts]
      responses:
        "200":
          description: Contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrustedContact"
    post:
      operationId: addTrustedContact
      tags: [trusted-contacts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrustedContactInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrustedContact"
  /trusted-contacts/{id}:
    delete:
      operationId: deleteTrustedContact
      tags: [trusted-contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Tracking ────────────────────────────────────────────────────────────────
  /trips/{token}:
    get:
      operationId: getPublicTrip
      tags: [tracking]
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public trip info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTrip"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Dashboard ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        demoToken:
          type: ["string", "null"]
        demoOtp:
          type: ["string", "null"]
...
    RideInput:
      type: object
      required: [origin, destination, date, time, vehicleType, vehicleNumber, fuelType, mileage, totalSeats, farePerSeat]
      properties:
        origin:
          type: string
        destination:
          type: string
        waypoints:
          type: array
          items:
            type: string
        date:
          type: string
        time:
          type: string
        vehicleType:
          type: string
          enum: [car, bike, auto, van]
        vehicleNumber:
          type: string
        fuelType:
          type: string
          enum: [petrol, diesel, electric, cng]
        mileage:
          type: number
        totalSeats:
          type: integer
        farePerSeat:
          type: number
        isRecurring:
          type: boolean
        recurringDays:
          type: array
          items:
            type: string
        eventId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    RideDetail:
      type: object
      required: [id, offererId, offererName, status, origin, destination, date, time, vehicleType, vehicleNumber, totalSeats, seatsAvailable, farePerSeat, createdAt]
      properties:
        id:
          type: integer
        offererId:
          type: integer
        offererName:
          type: string
        offererCollege:
          type: string
        offererVerified:
          type: boolean
        offererReliabilityScore:
          type: integer
        status:
          type: string
          enum: [open, booking_requested, accepted, passenger_picked_up, ride_started, completed, cancelled]
        origin:
          type: string
        destination:
          type: string
        waypoints:
          type: array
          items:
            type: string
        date:
          type: string
        time:
          type: string
        vehicleType:
          type: string
        vehicleNumber:
          type: string
        fuelType:
          type: string
        totalSeats:
          type: integer
        seatsAvailable:
          type: integer
        farePerSeat:
          type: number
        isRecurring:
          type: boolean
        notes:
          type: ["string", "null"]
        currentLat:
          type: ["number", "null"]
        currentLng:
          type: ["number", "null"]
        trackingToken:
          type: ["string", "null"]
        createdAt:
          type: string
    RideSearchResult:
      type: object
      required: [ride, matchScore, matchBreakdown]
      properties:
        ride:
          $ref: "#/components/schemas/RideDetail"
        matchScore:
          type: integer
        matchBreakdown:
          $ref: "#/components/schemas/MatchBreakdown"
    MatchBreakdown:
      type: object
      required: [destinationMatch, routeOverlap, timeMatch, pickupDistance, driverRating]
      properties:
        destinationMatch:
          type: integer
        routeOverlap:
          type: integer
        timeMatch:
          type: integer
        pickupDistance:
          type: integer
        driverRating:
          type: integer
    BookingInput:
      type: object
      required: [pickupPoint]
      properties:
        pickupPoint:
          type: string
        notes:
          type: ["string", "null"]
    Booking:
      type: object
      required: [id, rideId, passengerId, status, pickupPoint, otp, createdAt]
      properties:
        id:
          type: integer
        rideId:
          type: integer
        passengerId:
          type: integer
        passengerName:
          type: string
        status:
          type: string
          enum: [requested, accepted, rejected, picked_up, completed, cancelled]
        pickupPoint:
          type: string
        farePerSeat:
          type: number
        otp:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    BookingWithPassenger:
      type: object
      required: [id, rideId, passengerId, passengerName, passengerCollege, passengerVerified, status, pickupPoint, otp, createdAt]
      properties:
        id:
          type: integer
        rideId:
          type: integer
        passengerId:
          type: integer
        passengerName:
          type: string
        passengerCollege:
          type: string
        passengerVerified:
          type: boolean
        passengerReliabilityScore:
          type: integer
        passengerPhoto:
          type: ["string", "null"]
        status:
          type: string
        pickupPoint:
          type: string
        farePerSeat:
          type: number
        otp:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    LocationInput:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
...
    TrustedContact:
      type: object
      required: [id, userId, name, phone, relationship, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        relationship:
          type: string
          enum: [parent, guardian, friend, sibling, other]
        createdAt:
          type: string
    TrustedContactInput:
      type: object
      required: [name, phone, relationship]
      properties:
        name:
          type: string
        phone:
          type: string
        relationship:
          type: string
          enum: [parent, guardian, friend, sibling, other]

    # ─── Tracking ──────────────────────────────────────────────────────────────
    PublicTrip:
      type: object
      required: [rideId, driverName, vehicleType, vehicleNumber, origin, destination, status, createdAt]
      properties:
        rideId:
          type: integer
        driverName:
          type: string
        vehicleType:
          type: string
        vehicleNumber:
          type: string
        origin:
          type: string
        destination:
          type: string
        status:
          type: string
        currentLat:
          type: ["number", "null"]
        currentLng:
          type: ["number", "null"]
        eta:
          type: ["string", "null"]
        createdAt:
          type: string

    # ─── Dashboard ─────────────────────────────────────────────────────────────
...