# lib/api-zod/src/generated/types/trustedContactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ChaloTogether API specification
 * OpenAPI spec version: 0.2.0
 */
import type { TrustedContactInputRelationship } from './trustedContactInputRelationship';

export interface TrustedContactInput {
  name: string;
  phone: string;
  relationship: TrustedContactInputRelationship;
}